Move BitcoinApplication to header so it can be tested
Move-only commit, no code changes
